hyena-dna/requirements.txt
transformers
torchvision
einops
omegaconf
genomic-benchmarks
